examples/main: move to cmd/proxy and support as de-facto proxy binary.
